requirements.txt
confluent-kafka
gtfs-realtime-bindings
requests
pyspark==3.5.0
delta-spark==3.1.0
dotenv
apache-airflow
apache-airflow-providers-standard